+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { CartItemApi } from "@/apis";
 import { ShoppingBag } from "@/components/icons";
-import useQuery from "@/modules/Query/useQuery/useQuery";
+import useQuery from "@/modules/Query/useQuery";
 import Modal from "../Modal/Modal";
 import * as S from "./Header.styles";
 import ShoppingCartModal from "./ShoppingCartModal";
