--- src/modules/Query/QueryClient/QueryClient.ts
@@ -1,15 +1,0 @@
-export class QueryClient {
-  cache: Map<string, any> = new Map();
-
-  get(key: string) {
-    return this.cache.get(key);
-  }
-
-  set(key: string, value: any) {
-    this.cache.set(key, value);
-  }
-
-  delete(key: string) {
-    this.cache.delete(key);
-  }
-}
